tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/048.html

<!doctype html>
<title>document.write</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
onload = function() {
  t.step(function() {
    document.write("<body>Filler Text<div id='log'></div>");
    assert_equals(document.body.textContent, "Filler Text");
  });
  t.done();
};
</script>
<body>FAIL
<div id="log"></div>
